+++ agrest-jaxrs2-openapi/src/main/java/io/agrest/jaxrs2/openapi/modelconverter/AgEntityAwareModelConverter.java
@@ -1,6 +1,7 @@
 package io.agrest.jaxrs2.openapi.modelconverter;
 
 import io.agrest.PathConstants;
+import io.agrest.meta.AgAttribute;
 import io.agrest.meta.AgEntity;
 import io.agrest.meta.AgIdPart;
 import io.agrest.meta.AgRelationship;
@@ -30,53 +31,68 @@ public abstract class AgEntityAwareModelConverter extends AgModelConverter {
                 : context.resolve(new AnnotatedType().type(type));
     }
 
-    protected Schema doResolveId(AgEntity<?> entity, ModelConverterContext context) {
+    protected Schema doResolveId(AgEntity<?> entity, ModelConverterContext context, PropertyAccessChecker accessChecker) {
         switch (entity.getIdParts().size()) {
             case 0:
                 return null;
             case 1:
-                return doResolveSingleId(entity, context);
+                return doResolveSingleId(entity, context, accessChecker);
             default:
-                return doResolveIdMap(entity, context);
+                return doResolveIdMap(entity, context, accessChecker);
         }
     }
 
-    protected Schema doResolveSingleId(AgEntity<?> entity, ModelConverterContext context) {
+    protected Schema doResolveSingleId(AgEntity<?> entity, ModelConverterContext context, PropertyAccessChecker accessChecker) {
         AgIdPart idPart = entity.getIdParts().iterator().next();
-        return doResolveValue(idPart.getType(), context);
+        return accessChecker.allow(idPart) ? doResolveValue(idPart.getType(), context) : null;
     }
 
-    protected Schema doResolveIdMap(AgEntity<?> entity, ModelConverterContext context) {
+    protected Schema doResolveIdMap(AgEntity<?> entity, ModelConverterContext context, PropertyAccessChecker accessChecker) {
 
         Map<String, Schema> properties = new LinkedHashMap<>();
 
         List<AgIdPart> sortedIds = new ArrayList<>(entity.getIdParts());
         sortedIds.sort(Comparator.comparing(AgIdPart::getName));
         for (AgIdPart idPart : sortedIds) {
-            properties.put(idPart.getName(), doResolveValue(idPart.getType(), context));
+            if (accessChecker.allow(idPart)) {
+                properties.put(idPart.getName(), doResolveValue(idPart.getType(), context));
+            }
         }
 
-        return new ObjectSchema().properties(properties);
+        return properties.isEmpty() ? null : new ObjectSchema().properties(properties);
     }
 
-    protected Schema doResolveRelationship(AgRelationship relationship, ModelConverterContext context) {
+    protected Schema doResolveAttribute(AgAttribute attribute, ModelConverterContext context, PropertyAccessChecker accessChecker) {
+        return accessChecker.allow(attribute) ? doResolveValue(attribute.getType(), context) : null;
+    }
 
-        AgEntity<?> targetEntity = relationship.getTargetEntity();
+    protected Schema doResolveRelationship(AgRelationship r, ModelConverterContext context, PropertyAccessChecker accessChecker) {
+
+        if (!accessChecker.allow(r)) {
+            return null;
+        }
+
+        AgEntity<?> targetEntity = r.getTargetEntity();
 
         // ensure related entity and any other entities reachable from it are resolved
         context.resolve(new AnnotatedType().type(targetEntity.getType()));
 
         // link to resolved entity
         Schema relatedSchemaRef = new Schema().$ref(RefUtils.constructRef(targetEntity.getName()));
-        return relationship.isToMany()
+        return r.isToMany()
                 ? new ArraySchema().items(relatedSchemaRef)
                 : relatedSchemaRef;
     }
 
-    protected Schema doResolveRelationshipRef(AgRelationship r, ModelConverterContext context) {
+    protected Schema doResolveRelationshipRef(AgRelationship r, ModelConverterContext context, PropertyAccessChecker accessChecker) {
 
-        // relationships in updates are resolved as refs to IDs
-        Schema idSchema = doResolveId(r.getTargetEntity(), context);
+        if (!accessChecker.allow(r)) {
+            return null;
+        }
+
+        // relationships in updates are resolved as refs to IDs.. The permission here is for writing a relationship,
+        // not the ID, so since we just checked it above, always allow to proceed
+        Schema idSchema = doResolveId(r.getTargetEntity(), context, PropertyAccessChecker.allowAll());
         if (idSchema != null) {
             return r.isToMany() ? new ArraySchema().items(idSchema) : idSchema;
         }
@@ -84,13 +100,13 @@ public abstract class AgEntityAwareModelConverter extends AgModelConverter {
         return null;
     }
 
-    protected Map<String, Schema> doResolveProperties(
+    protected Map<String, Schema> doCollectProperties(
             Schema idSchema,
             List<Map.Entry<String, Schema>> attributesAndRelationships) {
 
         // property sorting should be stable and match that of DataResponse: "id" goes first, then a mix of
         // attributes and relationships in alphabetic order
-        
+
         Map<String, Schema> properties = new LinkedHashMap<>();
 
         if (idSchema != null) {